+++ tests/Reader/TestDataTest.php
@@ -51,6 +51,9 @@ class TestDataTest extends ParquetCsvComparison
    * [testAllTypesSnappyCompression description]
    */
   public function testAllTypesSnappyCompression(): void {
+    if(!extension_loaded('snappy')) {
+      static::markTestSkipped('ext-snappy unavailable');
+    }
     $this->CompareFiles('types/alltypes', 'snappy', true, [
       'integer',
       'boolean',